3027. Find the Number of Ways to Place People II

+++ Problems/3025. Find the Number of Ways to Place People I/3025_find_the_number_of_ways_to_place_people_i_cpp/main.cpp
@@ -42,7 +42,7 @@ int numberOfPairs(vector<vector<int>>& points) {
     return pairs;
 }
 
-void test(vector<vector<int>> points, double expected) {
+void test(vector<vector<int>> points, int expected) {
     cout << "Points: ";
     for(vector<int> point : points) {
         cout << "[" << point[0] << ", " << point[1] << "] ";
